Refactor eventsubscriber for eventseries/eventinstances into two seperate subscribers.
This seperates the logic for redirecting users/editors on eventseries/eventinstances into seperate subscribers.

This is attempt to ensure they both work as expected.

Code comments should explain throughly what they do, and why.

DDFFORM-475
